Rename and re-organize repository checks
