+++ lifemonitor/api/models/issues/general/repo_layout.py
@@ -24,30 +24,17 @@ import logging
 
 from lifemonitor.api.models.issues import WorkflowRepositoryIssue
 from lifemonitor.api.models.repositories import WorkflowRepository
+from .lm import MissingLMConfigFile
 
 # set module level logger
 logger = logging.getLogger(__name__)
 
 
-class MissingConfigFile(WorkflowRepositoryIssue):
-    name = "Missing config file"
-    description = "No <code>lifemonitor.yaml</code> configuration file found on this repository.<br>"\
-        "The <code>lifemonitor.yaml</code> should be placed on the root of this repository."
-    labels = ['config', 'enhancement']
-
-    def check(self, repo: WorkflowRepository) -> bool:
-        if repo.config is None:
-            config = repo.generate_config()
-            self.add_change(config)
-            return True
-        return False
-
-
-class NotInitialisedRepositoryIssue(WorkflowRepositoryIssue):
+class RepositoryNotInitialised(WorkflowRepositoryIssue):
     name = "Repository not intialised"
     description = "No workflow and crate metadata found on this repository."
     labels = ['invalid', 'enhancement', 'config']
-    depends_on = [MissingConfigFile]
+    depends_on = [MissingLMConfigFile]
 
     def check(self, repo: WorkflowRepository) -> bool:
         return repo.find_workflow() is None and repo.metadata is None
@@ -58,13 +45,13 @@ class MissingWorkflowFile(WorkflowRepositoryIssue):
     description = "No workflow found on this repository.<br>"\
         "You should place the workflow file (e.g., <code>.ga</code> file) according to the best practices ."
     labels = ['invalid', 'bug']
-    depends_on = [NotInitialisedRepositoryIssue]
+    depends_on = [RepositoryNotInitialised]
 
     def check(self, repo: WorkflowRepository) -> bool:
         return repo.find_workflow() is None
 
 
-class MissingMetadataFile(WorkflowRepositoryIssue):
+class MissingROCrateFile(WorkflowRepositoryIssue):
     name = "Missing RO-Crate metadata"
     description = "No <code>ro-crate-metadata.json</code> found on this repository.<br>"\
         "The <code>ro-crate-metadata.json</code> should be placed on the root of this repository."
@@ -79,11 +66,11 @@ class MissingMetadataFile(WorkflowRepositoryIssue):
         return False
 
 
-class MissingRoCrateWorkflowFile(WorkflowRepositoryIssue):
+class MissingROCrateWorkflowFile(WorkflowRepositoryIssue):
     name = "Missing RO-Crate workflow file"
     description = "The workflow file declared on RO-Crate metadata is missing in this repository."
     labels = ['invalid', 'bug']
-    depends_on = [MissingMetadataFile]
+    depends_on = [MissingROCrateFile]
 
     def check(self, repo: WorkflowRepository) -> bool:
         if repo.metadata:
